Use int for -X for UInt16
Because -((uint16_t)X) returns positive.

+++ lib/expr.c
@@ -2243,7 +2243,7 @@ grn_proc_call(grn_ctx *ctx, grn_obj *proc, int nargs, grn_obj *caller)
     break;                                                              \
   case GRN_DB_UINT16 :                                                  \
     {                                                                   \
-      uint16_t x_;                                                      \
+      int x_;                                                           \
       x_ = GRN_UINT16_VALUE(x);                                         \
       left_expression_check(x_);                                        \
       GRN_INT32_SET(ctx, res, integer_operation(x_));                   \
